Update CK Editor config
The CK Editor works perfectly in development, however does not work
properly in production. Issue fully documented in GH Issue #49.

The thinking is that it is something to do with the collection of static
files.

This commit adjusts the configuration in accordance with a suggestion in
the docs in an attempt to fix the issue.

GitHub Issue: #49