Rename project directory to `aa_project`
The project no longer has a Django back-end with a React front-end,
therefore there's no need for the main project directories to be named
as so.

`aa_project` is a great choice of name because it places the project
directory within my IDE in my preferred location (at the top of the
project structure).

Contributes towards: #57