Add Redis caching for scraping app
The scraping app, especially the scraping of the EU Referendum Results,
takes a long time to scrape from the BBC website.

This is required since it's a respectful process to not bombard the BBC
servers with repeated requests, however the content is static since
those referendum results will never need to be refreshed.

This change aims to deliver a better experience for the site's visitors.

Addresses the following:

- Implements indefinite caching for the scraped content views
- Adds the `heroku-redis` add-on to my account
- First attempt at its deployment
- Updates build process to only include production packages for Heroku's
  Dockerfile

Following the changes, all tests still pass.

Contributes towards: #60