Refactor referendum script
Addresses the following:

- Change the approach from a LBYL approach to an EAFP one
- Makes docstrings more concise
- Formatting changes to imports on `urls.py` file in `pages` app

Unless the BBC website were to ever return a server error or a forbidden
response, trying to get one of the pages will always succeed
irrespective of whether a `200` or `404` status code is returned. A `#
pragma: no cover` has been added to declare that I'm OK with not testing
that case within the test suite.

Ref: #40